+++ donators.py
@@ -1,10 +1,10 @@
-__version__ = (0, 0, 30)
+__version__ = (0, 0, 31)
 
 
 # ▄▀█ █▄ █ █▀█ █▄ █ █▀█ ▀▀█ █▀█ █ █ █▀
 # █▀█ █ ▀█ █▄█ █ ▀█ ▀▀█   █ ▀▀█ ▀▀█ ▄█
 #
-#           © Copyright 2023
+#           © Copyright 2024
 #
 #        developed by @anon97945
 #
@@ -15,8 +15,8 @@ __version__ = (0, 0, 30)
 # 🌐 https://www.gnu.org/licenses/gpl-3.0.html
 
 # meta developer: @apodiktum_modules
-# meta banner: https://t.me/file_dumbster/11
-# meta pic: https://t.me/file_dumbster/13
+# meta banner: https://t.me/apodiktum_dumpster/11
+# meta pic: https://t.me/apodiktum_dumpster/13
 
 # scope: hikka_only
 # scope: hikka_min 1.3.3
@@ -289,7 +289,7 @@ class ApodiktumDonatorsMod(loader.Module):
             int(self.config["logchannel"]),
             string_kick,
             schedule=(
-                date.today() + timedelta(days=(int(amount) / monthly_amount * 30))
+                date.today() + timedelta(days=int(amount) / monthly_amount * 30)
             ),
         )
         if self.config["custom_message"]: